Consolidate the Styles slots
